lib: split functions into separate files
