--- .npmignore
@@ -1,5 +1,0 @@
-# folders
-docs/
-test/
-examples/
-benchmarks/
